requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Training and evaluation
tqdm>=4.65.0
torch-fidelity>=0.3.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data and utilities
requests>=2.31.0
Pillow>=9.5.0
opencv-python>=4.7.0

# CLIP support 
transformers>=4.30.0
ftfy>=6.1.0
regex>=2023.0.0
# Install CLIP with: pip install git+https://github.com/openai/CLIP.git

# Logging and monitoring (optional)
tensorboard>=2.13.0
pyyaml>=6.0
wandb>=0.15.0  # For experiment tracking

# Code quality 
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0